Fix for Aliasing
https://github.com/mitchellh/goamz/blob/master/route53/route53.go#L101

According to Creating Alias Resource Record Sets documentation (http://docs.aws.amazon.com/Route53/latest/APIReference/CreateAliasRRSAPI.html), we do not include the XML element `<TTL>` when we want to create an alias but the present code inserts this element when the `query` is called by `ChangeResourceRecordSets`.

So in case we need to create an alias with load balancer, we need to take out `<TTL>` element. Kind of dirty but it works. Without this, I keep getting error of `client.ChangeResourceRecordSets error:Request failed, got status code: 400. Response: <?xml version="1.0"?>`