ec2: add InstancesPaginate and VolumesPaginate
The current API doesn't allow us to use MaxResults and NextToken. Those
are orthogonal to the current API (ids and filters can't be used).

This is needed so we can fetch thousands of Instances or Volumes in
batches. The normal API will not allow it, as it's going to time out
due to high latency.