Fix the EC2 block device mappings for image operations
Commit 9e419baaaa8185eb20a782ff34003478b683effb introduced a regression
that caused any calls for ephemeral disks or no device to include
`Ebs.DeleteOnTermination`, which caused AWS to think we were trying to
configure an EBS volume. Since the EBS volume didn't have a VolumeSize
or SnapshotId, this predictably failed.

This fixes the branching so when VirtualName or NoDevice are specified
it will not attempt to add the EBS related items.